changed 'meli_code' data type

+++ backend/database/migrations/0001_01_01_000000_create_users_table.php
@@ -18,7 +18,7 @@ return new class extends Migration
             $table->string('family')->nullable();
             $table->enum('role', ['user', 'supervisor', 'developer' ,'admin'])->default('user');
             $table->date('birth_date')->nullable();
-            $table->integer('meli_code')->nullable();
+            $table->bigint('meli_code')->nullable();
             $table->string('phone')->nullable();
             $table->enum('education', ['دیپلم', 'فوق دیپلم', 'لیسانس' ,'دکترا' , 'پرفسورا'])->nullable();;
             $table->string('address')->nullable();
